feat: separate into workspace + add idl converter (#6)
## Summary

Added a converter that attempts to extract a _classic_ IDL from a new
IDL format.

## Details

Separated into workspace to add converter crate.

Implemented best attempts for all parts of the IDL to convert, reporting
errors for cases that
cannot be resolved.

Added CI and a readme for the new crate + moved parts of the Readme to
the definition crate.